+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - improved message presentation of the multiline bash tool command
 - added validator script for extension UI component to be used by extension-creator skill
 - added responsive design for Settings
+- fixed providers API not returning data after update and added runtime array checks for provider data
 
 ## [0.79.0]